+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 Works with PHP 8.1.
 
 
+## v2.3
+
+Feature: Added `isEqualTo` and `isNotEqualTo` to `IsIntStringMapType`.
+Overall Mutation Code Coverage has risen from 97% to 98%. No value has decreased.
+
+Feature: Added `withValues`, `withoutValues` and `tryWithoutValues` to `IsCollectionType`, allowing to add/remove multiple values with a single method call. MSI has risen from 94% to 95% and Covered Code MSI from 96% to 97%.
+
+
 ### v2.2
 
 Feature: It is now possible to also transform values in `IsStringEnumType` before validating.